Handle asset spec base url with relative output
When the base url has been manually configured with an asset spec,
handle the case when the static route has been registered with an
asset spec as well. In this case, the base class resolver won't
properly prefix the url so it needs to manually be based off the
configured value. Either this derived path or the base class result
must then be resolved with `static_url` to finalize the result.