Pass un-molested settings to add_static_view
Pyramid users expect to be able to register static views with asset
specs and then generate static urls with asset specs. Since webassets
needs the spec to be resolved before we pass it the configuration, we
should pass the original settings to the `add_static_view` call so that
if 'base_dir' is an asset spec 'static_url' works as expected.