Reconstruct globs when input and base are specs
When the base directory is an asset spec and an input file is also
an asset spec that contains a glob then the static view machinery
will be called with the glob.

In the past, this was not a problem because the `static_url` call
would always fail on the spec. Now that an asset spec in the
`webassets.base_dir` setting is honored with `webassets.static_view`
URLs containing a glob are returned instead of failing and falling
back to resolution using the filepath that was expanded during the
search for the source file.