With static_view add the externals separately
When `webassets.static_view` is used and the base directory is an asset
specification the `resolve_source_to_url` method of the resolver will
receive an absolute path to the pulled file and the original asset spec
for the input but has no access to the original base directory setting
in order to construct a new specification to pass to `static_url`. By
adding an additional static view for the externals using the absolute
path of the external directory these paths will resolve using the
Pyramid `static_url` machinery.