Clean up static view asset spec tests
These tests were somewhat confusing. They referred to base_url a lot
but actually meant to be setting the base_dir. Additionally, some
tests that seemed to pass should have been failing because they were
referencing non-existent files.

- Add the `static:some` temp files
- Change base_url parameterization to base_dir

This change causes a test to fail, which reveals an edge case in how
output URLs are handled. Relative outputs with a base directory as an
asset specification will not be resolved with static_url.