Handle multiple bundles with overrides better
The YAML parser will override declarations if they are redeclared.
In order to allow one YAML bundle definition file to override named
bundles in others all the files must be parsed as one YAML stream.
It is not sufficient to pass the existing environment to the loader
because even if the environment let us declare a bundle twice the
earlier inter-bundle references might be incorrect. Parsing all the
files as one stream solves the problem.